remove generators
